+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
-## [v3.0.0-alpha1] - 25 March 2021
+## [v3.0.0-beta1] - 10 September 2021
+
+* Remove Microservice output from `get all` command
+* Fix Debian Stretch Docker installation
+* Fix Debian Buster apt update release info issue for Agent install
+
+## [v3.0.0-alpha1] - 11 March 2021
 
 * Add Application Template support
 * Add Agent `upgrade` and `rollback` commands